+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - AS1/2 - highlight variable definition and all its instances on cursor place
   (also in edit mode)
 - AS1/2 - underline errors in the code (also in edit mode) 
+- AS1/2 - highlight variables and errors on scrollbar
 
 ### Changed
 - AS1/2 - Single DoAction tag inside frame is now displayed directly as frame node